+++ clientsrc/src/components/TaskComp.vue
@@ -1,31 +1,49 @@
 <template>
   <div class="task-comp">
-      <li class="d-flex justify-content-between">
-        <select
-          v-model="selectListData.listId"
-          @change="selectList"
-          class="form-control form-control-sm"
-        >
-          <option :value="list.id" v-for="list in lists" :key="list.id">{{list.title}}</option>
-        </select>
-        <p class="font-weight-bold">{{taskProp.description}}</p>
-        <div>
-        <i class="fa fa-angle-up mr-4 text-info" @click="commentsToggle=!commentsToggle" v-if="commentsToggle"></i>
-        <i class="fa fa-angle-down mr-4 text-info" @click="commentsToggle=!commentsToggle" v-if="!commentsToggle"></i>
+    <li class="d-flex justify-content-between">
+      <select
+        v-model="selectListData.listId"
+        @change="selectList"
+        class="form-control form-control-sm"
+      >
+        <option :value="list.id" v-for="list in lists" :key="list.id">
+          {{ list.title }}
+        </option>
+      </select>
+      <p class="font-weight-bold">{{ taskProp.description }}</p>
+      <div>
+        <i
+          class="fa fa-angle-up mr-4 text-info"
+          @click="commentsToggle = !commentsToggle"
+          v-if="commentsToggle"
+        ></i>
+        <i
+          class="fa fa-angle-down mr-4 text-info"
+          @click="commentsToggle = !commentsToggle"
+          v-if="!commentsToggle"
+        ></i>
         <i class="fa fa-trash text-danger" @click="deleteTask"></i>
-        </div>
-      </li>
+      </div>
+    </li>
 
     <ul v-if="commentsToggle">
-      <li class="d-flex justify-content-start" v-for="(comment,index) in taskProp.comments" :key="comment.id">
+      <li
+        class="d-flex justify-content-start"
+        v-for="(comment, index) in taskProp.comments"
+        :key="comment.id"
+      >
         <i class="fa fa-minus mr-2" @click="deleteComment(index)"></i>
         <p class="text-muted">
-        {{comment.description}}
+          {{ comment.description }}
         </p>
       </li>
     </ul>
 
-    <form @submit.prevent="addComment" class="d-flex justify-content-end" v-if="commentsToggle">
+    <form
+      @submit.prevent="addComment"
+      class="d-flex justify-content-end"
+      v-if="commentsToggle"
+    >
       <div class="form-group form-inline">
         <input
           type="text"
@@ -33,7 +51,7 @@
           placeholder="Add Comment..."
           v-model="addCommentData.description"
         />
-      <button class="btn-sm btn-warning" type="submit">Add Comment</button>
+        <button class="btn-sm btn-warning" type="submit">Add Comment</button>
       </div>
     </form>
   </div>
@@ -41,6 +59,7 @@
 
 
 <script>
+import sweetAlert from "../Utils/sweetAlert2.js";
 export default {
   name: "task-comp",
   props: ["taskProp", "index"],
@@ -48,7 +67,7 @@ export default {
     return {
       selectListData: {},
       addCommentData: {},
-      commentsToggle: false
+      commentsToggle: false,
     };
   },
   computed: {
@@ -57,8 +76,9 @@ export default {
     },
   },
   methods: {
-    deleteTask() {
-      this.$store.dispatch("deleteTask", this.taskProp);
+    async deleteTask() {
+      if (await sweetAlert.confirmAction())
+        this.$store.dispatch("deleteTask", this.taskProp);
     },
     selectList() {
       this.$store.dispatch("selectList", {
@@ -90,8 +110,7 @@ select {
   width: 0;
   cursor: pointer;
 }
-i{
+i {
   cursor: pointer;
 }
-
 </style>
